pyproject.toml
[tool.poetry]
name = "restaurant-rag-bot1"
version = "0.1.0"
description = "A LINE Bot powered by openai and pinecone for restaurant staff Q&A"
authors = ["anton_shih"]
package-mode = false


[tool.poetry.dependencies]
python = ">=3.10,<3.14"

# openai SDK
openai = "^1.77.0"

# 向量資料庫
pinecone = "^6.0.2"

# Web 後端
fastapi = "^0.111.0"
uvicorn = "^0.29.0"

# LINE Bot
line-bot-sdk = "^3.14.2"

# 常用工具
pandas = "^2.2.3"
pydantic = "^2.10.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

# 測試
pytest = "^8.3.4"
pillow = "^11.2.1"
motor = "^3.7.1"
pymongo = "^4.13.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


